+++ lib/dead_end.rb
@@ -1,4 +1,3 @@
 # frozen_string_literal: true
 
-require_relative "dead_end/api"
 require_relative "dead_end/core_ext"